+++ Assets/Plugins/AnimTransferer/Scripts/Runtime/Core/AnimTransferGroup.cs
@@ -11,11 +11,15 @@ namespace AnimTransferer.Runtime
         public string groupName;
         [Tooltip("When enable, exported json will include anim asset's file path & guid as path info. Allow editor load asset back from these path when loading json data")]
         public bool exportAnimAssetData;
-        [Tooltip("When enable, AnimTransferer will find states in the target controller with similar name compared to the Clip-State-Pair and transfer motion into it")]
+        [Tooltip("When enable, Clip that in source animations with similar name compared to the config of Clip-State-Pairs will also seen as transferable Clip")]
+        public bool findSimilarClipName;
+        [DrawIf("findSimilarClipName", true), Tooltip("When the similarity between the name of Clip and any Source Clip Name in Clip-State-Pairs is greater or equal to the set value, the editor will treat it as a transferable Clip")]
+        public float preferedSimilarityOfClipName;
+        [Tooltip("When enable, State that in the target controller with similar name compared to the config of Clip-State-Pairs will also seen as transferable State")]
         public bool findSimilarStateName;
-        [DrawIf("findSimilarStateName", true), Tooltip("When the similarity between the name of State in the animator controller and Clip-State-Pair is greater than the set Similarity value, the editor will treat it as a replaceable State")] 
-        public float similarity;
-        [DrawIf("findSimilarStateName", true), Tooltip("When enable, AnimTransferer will also replace matched state name from config")] 
+        [DrawIf("findSimilarStateName", true), Tooltip("When the similarity between the name of State and any Target Animator Controller Name in Clip-State-Pairs is greater or equal to the set value, the editor will treat it as a transferable State")]
+        public float preferedSimilarityOfStateName;
+        [DrawIf("findSimilarStateName", true), Tooltip("When enable, matched State's name will replace with most similar name found in the config of Clip-State-Pairs")]
         public bool replaceMatchedStateNameWithConfig;
         [Tooltip("AnimTransferer will try finding clips in SourceAnimationClips in AnimAssets with SourceClipName,and then, find matched state name in the TargetAnimatorController in AnimAssets with TargetStateName,finally,replace matched state's motion's clip with matched clip")]
         public List<ClipStatePair> clipStatePairs;