+++ Assets/Plugins/AnimTransferer/CHANGELOG.md
@@ -1,5 +1,10 @@
 ## CHANGELOG
 
+### v1.1.0
+#### Modification of AnimTransferGroup :
+1. Modified variable name and tooltip.
+2. Added variable **findSimilarClipName(bool)** and **preferedSimilarityOfClipName(float)**, work exactly like **findSimilarStateName(bool)** and **preferedSimilarityOfStateName(float)** but it's for clips that in source animations to compare with config of Clip-State-Pairs while transfering.
+
 ### v1.0.1
 - Added tooltip to the AnimTransferGroup.